+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - ir: Add various `unit()`, `unit_mut()`, `get_unit()`, and `get_unit_mut()` functions.
 
+### Fixed
+- llhd-check: Honor dump flag `-d`.
+
 ## 0.10.0 - 2019-11-24
 ### Added
 - Add Sublime Text syntax highlighting.